vralign automatic

+++ projects/DLC_behavior_classification/preprocessing.py
@@ -60,6 +60,20 @@ def match_eye_to_tail_videos(eyevids,dst, vidpth = r'\\storage1.ris.wustl.edu\eb
             
     return vids2get
 
+def consecutive_stretch(x):
+    z = np.diff(x)
+    break_point = np.where(z != 1)[0]
+
+    if len(break_point) == 0:
+        return [x]
+
+    y = [x[:break_point[0]]]
+    for i in range(1, len(break_point)):
+        y.append(x[break_point[i - 1] + 1:break_point[i]])
+    y.append(x[break_point[-1] + 1:])
+    
+    return y 
+
 def copyvr_dlc(vrdir, dlcfls): #TODO: find a way to do the same for clampex
     """copies vr files for existing dlc csvs
     essentially do not put a csv into this folder if you don't want it analysed
@@ -103,7 +117,7 @@ def copyvr_dlc(vrdir, dlcfls): #TODO: find a way to do the same for clampex
             datetime_object = datetime.strptime(s[0][0], '%d_%b_%Y')                
             dt = str(datetime_object.date())
             if 'test'.lower() not in xx and 'test'.upper() not in xx and dt in dates:
-                shutil.copy(xx,dlcfls)
+                if not os.path.exists(os.path.join(dlcfls, os.path.basename(xx))): shutil.copy(xx,dlcfls)
                 mouse_vr.append([mouse, xx, dt])
                 
         print(f"\n********* copied vr files to dlc pose data for {mouse} *********")
@@ -168,7 +182,9 @@ def hdf5_to_dict(hdf5_object):
 
 def VRalign_automatic(vrfl, dlccsv, savedst, only_add_experiment=False):
     """zahra's implementation for VRstandendsplit for python dlc pipeline
-
+    automatic alignment
+    saves a png file with the imagesync start and stop so in the end you can corroborate
+    and make sure recording doesnt have multiple imagesync chunks
     Args:
         vrfl (_type_): _description_
         dlccsv (_type_): _description_
@@ -202,8 +218,8 @@ def VRalign_automatic(vrfl, dlccsv, savedst, only_add_experiment=False):
             # assumes there are no additional images after behavior
             # also prints out the imagesync plots if you want to do it manually?
             iinds = np.where(imageSync>1)
-            uscanstart = np.min(iinds)+1
-            uscanstop  = np.max(iinds)+1
+            uscanstart = np.min(iinds)
+            uscanstop  = np.max(iinds)
             print(f'Start of scan: {uscanstart}')
             print(f'End of scan: {uscanstop}')
             print(f'Length of scan is {uscanstop - uscanstart}')
@@ -232,11 +248,17 @@ def VRalign_automatic(vrfl, dlccsv, savedst, only_add_experiment=False):
             #imaging data from before you started VR so sets to 0
 
             # cuts all of the variables from VR
+            # zahra's hack to fix vr variables after interpolation
             urewards = np.squeeze(VR['reward'][scanstart:scanstop]); 
             urewards_cs = np.zeros_like(urewards)
             urewards_cs = urewards==0.5
-            urewards_cs = interpolate_vrdata(uscanstop,uscanstart,dlcdf,urewards_cs)
+            urewards_cs = interpolate_vrdata(uscanstop,uscanstart,dlcdf,urewards_cs)            
             urewards=urewards_cs>0 # doesn't have reward variable anymore, so add that if needed (after 500msec)
+            # fix multiple frames per cs
+            urewards_ = consecutive_stretch(np.where(urewards>0)[0])
+            urewards_ = np.array([min(xx) for xx in urewards_])
+            urewards = np.zeros_like(urewards)
+            urewards[urewards_]=1 # assign cs to boolean
             uimageSync = np.squeeze(VR['imageSync'][scanstart:scanstop]); uimageSync = interpolate_vrdata(uscanstop,uscanstart,dlcdf,uimageSync); 
             uforwardvel = np.hstack(-0.013*VR['ROE'][scanstart:scanstop])/np.diff(np.squeeze(VR['time'][scanstart-1:scanstop]))
             uforwardvel = interpolate_vrdata(uscanstop,uscanstart,dlcdf,uforwardvel)
